backend/requirements.txt
fastapi
uvicorn
yt-dlp
faster-whisper
ffmpeg-python
python-dotenv
pydantic
google-generativeai
Pillow
sqlmodel
